Move all extension tests into test sub-directories
