+++ src/extensions/lightbox-controls/test/lightbox-controls.cypress.js
@@ -1,7 +1,7 @@
 /*
  * Include our constants
  */
-import * as helpers from '../../../.dev/tests/cypress/helpers';
+import * as helpers from '../../../../.dev/tests/cypress/helpers';
 
 describe( 'Test CoBlocks Lightbox Controls extension', function() {
 	/**
